+++ server/esSearch.js
@@ -702,12 +702,12 @@ Meteor.methods({
           }
           search_query.body["sort"]=[{"s":"asc"},{"a":"asc"}]
           search_query.body.highlight={}
-          search_query.body.suggest={}
+          // search_query.body.suggest={}
           search_query.body.min_score=1
         } else {
           search_query.body.query.bool["should"] = shouldDSL
           search_query.body.query.bool["minimum_should_match"] = 1
-          search_query.body.highlight={}
+          // search_query.body.highlight={}
         }
       }
       // console.log(JSON.stringify(search_query.body.from));