+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 6.41:
+  * Enabling highlights for filter queries as well
+
 ## 6.4:
   * Aggregate section updated to allow layer and sublayer filtering in the query
   * Suggest support for query layer and sublayer filtering added